Fix ClassCastException when the REPL is launched

+++ src/org/jetbrains/plugins/clojure/repl/impl/ClojureProcessREPL.java
@@ -102,7 +102,7 @@ public class ClojureProcessREPL extends REPLBase
     processHandler.startNotify();
 
     SafeFn waitForAck = SafeFn.find(REPLComponent.NREPL_NS, WAIT_FOR_ACK);
-    Integer maybePort = (Integer) waitForAck.sInvoke(Long.valueOf(PROCESS_WAIT_TIME));
+    Long maybePort = (Long) waitForAck.sInvoke(Long.valueOf(PROCESS_WAIT_TIME));
 
     if (maybePort == null)
     {
